+++ src/lib/qs/array.ts
@@ -1,4 +1,6 @@
+import { createInvalidArrayItemsError } from "../errors/invalid-array-items";
 import { createInvalidTypeError } from "../errors/invalid-type";
+import { createMaxArrayLengthError } from "../errors/max-array-length";
 import { Serializer } from "../types";
 import { ArrayType, name as typeName } from "../types/array";
 
@@ -8,24 +10,39 @@ export function register(serializer: Serializer): void {
   }
 
   function read<T>(type: ArrayType<T>, input: any[] | undefined): T[] {
-    let result: T[];
-    if (Array.isArray(input)) {
-      result = input.map((item: any): T => serializer.read(type.itemType, item));
-    } else if (input === undefined) {
-      result = [];
-    } else {
+    if (input === undefined) {
+      return [];
+    }
+    if (!Array.isArray(input)) {
       throw createInvalidTypeError("array | undefined", input);
     }
-    const error: Error | undefined = type.testError(result);
-    if (error !== undefined) {
-      throw error;
+    if (type.maxLength !== undefined && input.length > type.maxLength) {
+      throw createMaxArrayLengthError(input, type.maxLength);
+    }
+    let invalid: undefined | Map<number, Error> = undefined;
+    const result: T[] = [];
+    const itemCount: number = input.length;
+    for (let i: number = 0; i < itemCount; i++) {
+      try {
+        const item: T = serializer.read(type.itemType, input[i]);
+        if (invalid === undefined) {
+          result.push(item);
+        }
+      } catch (err) {
+        if (invalid === undefined) {
+          invalid = new Map();
+        }
+        invalid.set(i, err);
+      }
+    }
+    if (invalid !== undefined) {
+      throw createInvalidArrayItemsError(invalid);
     }
     return result;
   }
 
   function readTrusted<T>(type: ArrayType<T>, input: any[] | undefined): T[] {
     if (Array.isArray(input)) {
-      // TODO(demurgos): Avoid casting
       return input.map((item: any): T => serializer.readTrusted(type.itemType, item));
     } else {
       return [];